+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @longo-ignite-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- Add Toast component and changes in tokens library
+
 ## 2.0.0
 
 ### Major Changes