+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @longo-ignite-ui/tokens
 
+## 2.0.0
+
+### Major Changes
+
+- Add Toast component and changes in tokens library
+
 ## 1.1.0
 
 ### Minor Changes